+++ neutrino2/src/gui/widget/infobox.h
@@ -114,7 +114,7 @@ class CInfoBox
 		bool setText(const char * const newText, const char * const _thumbnail = NULL, int _tw = 0, int _th = 0, int tmode = CTextBox::PIC_RIGHT, bool enable_frame = false, const bool useBackground = false);
 		void setMode(const int mode){m_nMode = mode;};
 		void setBackgroundColor(fb_pixel_t col);
-		void setTextColor(uint8_t col);
+		void setTextColor(uint32_t col);
 		void setFont(unsigned int font_text);
 		void setBorderMode(int sm = CComponent::BORDER_ALL){borderMode = sm;};
